requirements.txt
flask
psycopg2-binary
requests
gunicorn
ftfy
bs4==0.0.1
lxml==4.5.0
testing.postgresql